# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NestyAI",
    "description": "Personal AI Gateway Server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/internal/model-configs/{model_id}": {
      "get": {
        "tags": [
          "internal-model-configs"
        ],
        "summary": "Get Internal Model Config",
        "operationId": "get_internal_model_config_internal_model_configs__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Internal Model Config Internal Model Configs  Model Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "internal-model-configs"
        ],
        "summary": "Patch Internal Model Config",
        "operationId": "patch_internal_model_config_internal_model_configs__model_id__patch",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelOverridePatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Patch Internal Model Config Internal Model Configs  Model Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/model-configs/{model_id}/reset": {
      "post": {
        "tags": [
          "internal-model-configs"
        ],
        "summary": "Reset Internal Model Config",
        "operationId": "reset_internal_model_config_internal_model_configs__model_id__reset_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "changed_by_label",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "internal-admin",
              "title": "Changed By Label"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Reset Internal Model Config Internal Model Configs  Model Id  Reset Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/console/runtime/validate": {
      "post": {
        "tags": [
          "internal-console-runtime"
        ],
        "summary": "Runtime Validate",
        "operationId": "runtime_validate_internal_console_runtime_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Runtime Validate Internal Console Runtime Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/console/runtime/model-configs/{model_id}": {
      "post": {
        "tags": [
          "internal-console-runtime"
        ],
        "summary": "Runtime Update Model Config",
        "operationId": "runtime_update_model_config_internal_console_runtime_model_configs__model_id__post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeModelConfigUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Runtime Update Model Config Internal Console Runtime Model Configs  Model Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/console/runtime/model-configs/{model_id}/reset": {
      "post": {
        "tags": [
          "internal-console-runtime"
        ],
        "summary": "Runtime Reset Model Config",
        "operationId": "runtime_reset_model_config_internal_console_runtime_model_configs__model_id__reset_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "changed_by_label",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "internal-console",
              "title": "Changed By Label"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Runtime Reset Model Config Internal Console Runtime Model Configs  Model Id  Reset Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/console/runtime/provider-chain/{model_id}": {
      "post": {
        "tags": [
          "internal-console-runtime"
        ],
        "summary": "Runtime Update Provider Chain",
        "operationId": "runtime_update_provider_chain_internal_console_runtime_provider_chain__model_id__post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeProviderChainUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Runtime Update Provider Chain Internal Console Runtime Provider Chain  Model Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelOverridePatchRequest": {
        "properties": {
          "override": {
            "type": "object",
            "title": "Override"
          },
          "changed_by_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed By Label",
            "default": "internal-admin"
          }
        },
        "type": "object",
        "required": [
          "override"
        ],
        "title": "ModelOverridePatchRequest"
      },
...
      "RuntimeModelConfigUpdateRequest": {
        "properties": {
          "override": {
            "type": "object",
            "title": "Override"
          },
          "changed_by_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed By Label",
            "default": "internal-console"
          }
        },
        "type": "object",
        "required": [
          "override"
        ],
        "title": "RuntimeModelConfigUpdateRequest"
      },
...
      "RuntimeProviderChainUpdateRequest": {
        "properties": {
          "provider_chain": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Provider Chain"
          },
          "changed_by_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed By Label",
            "default": "internal-console"
          }
        },
        "type": "object",
        "required": [
          "provider_chain"
        ],
        "title": "RuntimeProviderChainUpdateRequest"
      },
...
      "RuntimeValidateRequest": {
        "properties": {
          "model_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Model Id"
          },
          "override": {
            "type": "object",
            "title": "Override"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "override"
        ],
        "title": "RuntimeValidateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}